+++ v3.x.x/main/autopatcher/Mods.cs
@@ -1,0 +1,14 @@
+﻿namespace Azurlane
+{
+    internal enum Mods
+    {
+        GodMode,
+        WeakEnemy,
+        GodMode_Damage,
+        GodMode_Cooldown,
+        GodMode_WeakEnemy,
+        GodMode_Damage_Cooldown,
+        GodMode_Damage_WeakEnemy,
+        GodMode_Damage_Cooldown_WeakEnemy
+    }
+}
