+++ v3.x.x/main/autopatcher/Program.cs
@@ -11,18 +11,6 @@ using Azurlane.Properties;
 
 namespace Azurlane
 {
-    internal enum Mods
-    {
-        GodMode,
-        WeakEnemy,
-        GodMode_Damage,
-        GodMode_Cooldown,
-        GodMode_WeakEnemy,
-        GodMode_Damage_Cooldown,
-        GodMode_Damage_WeakEnemy,
-        GodMode_Damage_Cooldown_WeakEnemy
-    }
-
     internal static class Program
     {
         internal static bool Abort;
