+++ v3.x.x/main/cli/OptionException.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.Runtime.Serialization;
+using System.Security.Permissions;
+
+namespace Azurlane
+{
+    [Serializable]
+    public class OptionException : Exception
+    {
+        public OptionException()
+        {
+        }
+
+        public OptionException(string message, string optionName)
+            : base(message)
+        {
+            OptionName = optionName;
+        }
+
+        public OptionException(string message, string optionName, Exception innerException)
+            : base(message, innerException)
+        {
+            OptionName = optionName;
+        }
+
+        protected OptionException(SerializationInfo info, StreamingContext context)
+            : base(info, context)
+        {
+            OptionName = info.GetString("OptionName");
+        }
+
+        public string OptionName { get; }
+
+        [SecurityPermission(SecurityAction.LinkDemand, SerializationFormatter = true)]
+        public override void GetObjectData(SerializationInfo info, StreamingContext context)
+        {
+            base.GetObjectData(info, context);
+            info.AddValue("OptionName", OptionName);
+        }
+    }
+}
