+++ v3.x.x/main/cli/OptionValueType.cs
@@ -1,0 +1,9 @@
+﻿namespace Azurlane
+{
+    public enum OptionValueType
+    {
+        None,
+        Optional,
+        Required
+    }
+}
