+++ v3.x.x/main/cli/Tasks.cs
@@ -1,0 +1,12 @@
+﻿namespace Azurlane
+{
+    internal enum Tasks
+    {
+        Encrypt,
+        Decrypt,
+        Decompile,
+        Recompile,
+        Unpack,
+        Repack
+    }
+}
