+++ v3.x.x/main/cli64/OptionAction.cs
@@ -1,0 +1,4 @@
+﻿namespace Azurlane
+{
+    public delegate void OptionAction<TKey, TValue>(TKey key, TValue value);
+}
