+++ v3.x.x/main/rewriter/IniFileParser/Model/Configuration/DefaultIniParserConfiguration.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace Azurlane.IniFileParser.Model.Configuration
+{
+    [Obsolete("Kept for backward compatibility, just use IniParserConfiguration class")]
+    public class DefaultIniParserConfiguration : ConcatenateDuplicatedKeysIniParserConfiguration
+    {
+    }
+}
