+++ v3.x.x/main/rewriter/IniFileParser/Model/Configuration/IniParserConfiguration.cs
@@ -4,11 +4,6 @@ using Azurlane.IniFileParser.Parser;
 
 namespace Azurlane.IniFileParser.Model.Configuration
 {
-    [Obsolete("Kept for backward compatibility, just use IniParserConfiguration class")]
-    public class DefaultIniParserConfiguration : ConcatenateDuplicatedKeysIniParserConfiguration
-    {
-    }
-
     /// <summary>
     ///     Defines data for a Parser configuration object.
     /// </summary>
